Move RoslynHost and Solutions
