+++ Source/Kesmai.WorldForge/Editor/MVP/Models/World/Segment.cs
@@ -10,11 +10,6 @@ using Kesmai.WorldForge.Models;
 using Kesmai.WorldForge.Scripting;
 using Kesmai.WorldForge.UI.Documents;
 using CommunityToolkit.Mvvm.ComponentModel;
-using Kesmai.WorldForge.Roslyn;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.Text;
-using RoslynPad.Roslyn;
 
 namespace Kesmai.WorldForge.Editor;
 
@@ -28,17 +23,11 @@ public class Segment : ObservableObject
 	private string _name;
 	private Script _internal;
 	private Script _definition;
-
-	private CustomRoslynHost _roslynHost;
-
+	
 	public string Name
 	{
 		get => _name;
-		set
-		{
-			if (SetProperty(ref _name, value))
-				UpdateWorkspace();
-		}
+		set => SetProperty(ref _name, value);
 	}
 
 	public Script Internal
@@ -52,9 +41,7 @@ public class Segment : ObservableObject
 		get => _definition;
 		set => SetProperty(ref _definition, value);
 	}
-	
-	public RoslynHost Roslyn => _roslynHost;
-	
+
 	public NotifyingCollection<SegmentRegion> Regions { get; set; } = new NotifyingCollection<SegmentRegion>();
 	public SegmentLocations Locations { get; set; } = new SegmentLocations();
 	public SegmentSubregions Subregions { get; set; } = new SegmentSubregions();
@@ -119,57 +106,6 @@ public class Segment : ObservableObject
 
 		if (updateOld)
 			args.OldItems.ForEach(region => { presenter.Documents.Remove(region); });
-	}
-
-	public void UpdateWorkspace()
-	{
-		var metadataReferences = AppDomain.CurrentDomain.GetAssemblies()
-			.Where(a => !a.IsDynamic && !String.IsNullOrEmpty(a.Location))
-			.Select(a => MetadataReference.CreateFromFile(a.Location))
-			.ToList();
-		
-		var scriptingData = Core.ScriptingData;
-        
-		if (scriptingData != null)
-			metadataReferences.Add(MetadataReference.CreateFromImage(scriptingData));
-
-		var serviceAssemblies = new[]
-		{
-			Assembly.Load("RoslynPad.Roslyn.Windows"),
-			Assembly.Load("RoslynPad.Editor.Windows")
-		};
-
-		var namespaceImports = new string[]
-		{
-			$"static Kesmai.Server.Segments.{Name}",
-		};
-		
-		var roslynReferences = RoslynHostReferences.NamespaceDefault
-			.With(references: metadataReferences, imports: namespaceImports);
-		
-		_roslynHost = new CustomRoslynHost(serviceAssemblies, roslynReferences);
-
-		var workspace = _roslynHost.Workspace;
-		var solution = workspace.CurrentSolution;
-		
-		var project = solution.AddProject($"Segment", $"Kesmai.Server.Segments.{Name}", LanguageNames.CSharp)
-			/* C# minimum to support global usings. */
-			.WithParseOptions(new CSharpParseOptions(LanguageVersion.CSharp10))
-			/* Minimum references to prevent overloading */
-			.AddMetadataReference(MetadataReference.CreateFromFile(typeof(object).Assembly.Location))
-			.AddMetadataReference(MetadataReference.CreateFromFile(typeof(Enumerable).Assembly.Location));
-
-		solution = project.Solution;
-
-		var directory = new DirectoryInfo(@"C:\Example\Source");
-		
-		foreach(var file in directory.GetFiles("*.cs", SearchOption.AllDirectories))
-		{
-			solution = solution.AddDocument(DocumentId.CreateNewId(project.Id), file.Name, 
-				SourceText.From(File.ReadAllText(file.FullName)));
-		}
-		
-		workspace.TryApplyChanges(solution);
 	}
 	
 	public void Load(XElement element)
