+++ Source/Kesmai.WorldForge/Editor/Scripting/ScriptEditor.cs
@@ -75,9 +75,8 @@ public class ScriptEditor : RoslynCodeEditor
 
         var services = (ServiceContainer)ServiceLocator.Current;
         var presenter = services.GetInstance<ApplicationPresenter>();
-        var segment = presenter.Segment;
 
-        await InitializeAsync(segment.Roslyn, new ClassificationHighlightColors(),
+        await InitializeAsync(presenter.Roslyn, new ClassificationHighlightColors(),
             Directory.GetCurrentDirectory(), string.Empty, SourceCodeKind.Regular);
 
         // Delay folding until TextArea is fully initialized
